PL-9314 - Add ResponseHelper.put(JSONObject o, String key, Object value) to replace the many methods with the same signature in Renderers. PL-1764 - Enable tag deletion in the tag resource. PL-1766 - Enforce permissions for the tag resource. PL-1979 - Suppress tag count when it's not required. PL-1765 - Fuller representation for tags resource.

+++ src/main/java/com/amee/platform/resource/tag/v_3_2/TagRemover_3_2_0.java
@@ -7,6 +7,7 @@ import com.amee.base.resource.RequestWrapper;
 import com.amee.base.resource.ResponseHelper;
 import com.amee.domain.tag.Tag;
 import com.amee.platform.resource.tag.TagResource;
+import com.amee.service.auth.ResourceAuthorizationService;
 import com.amee.service.tag.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
@@ -20,11 +21,19 @@ public class TagRemover_3_2_0 implements TagResource.Remover {
     @Autowired
     private TagService tagService;
 
+    @Autowired
+    private ResourceAuthorizationService resourceAuthorizationService;
+
+    @Override
     public Object handle(RequestWrapper requestWrapper) {
         String tagIdentifier = requestWrapper.getAttributes().get("tagIdentifier");
         if (tagIdentifier != null) {
             Tag tag = tagService.getTagByIdentifier(tagIdentifier);
             if (tag != null) {
+                // Authorized?
+                resourceAuthorizationService.ensureAuthorizedForRemove(
+                        requestWrapper.getAttributes().get("activeUserUid"), tag);
+                // Handle ReturnValueDefinition removal.
                 tagService.remove(tag);
                 return ResponseHelper.getOK(requestWrapper);
             } else {
